+++ docs/changes.rst
@@ -4,6 +4,10 @@ Changelog
 1.0 - Unreleased
 ----------------
 
+* Add `plone.app.testing <http://pypi.python.org/pypi/plone.app.testing>`_
+  based test fixture.
+  [wichert]
+
 * Delete some old copy/paste leftovers from `Euphorie
   <http://pypi.python.org/pypi/Euphorie>`_.
   [wichert]